requirements.txt
requests~=2.32.4
beautifulsoup4~=4.13.4
python-dateutil~=2.9.0.post0